requirements.txt
python-dotenv
pyrogram
